1. Make compile with clang++ 3.1 2. Add the sparc feature to disable sparc (eigen 3.0.5's optimization module fails with clang currently) 3. Make read-only attribute templated, so that serialization code can be generated at compile-time and non-serializable types are OK 4. Remove Se3 from Renderer code in favor of pos+ori separately (Se3 can be removed completely in the future) 5. Fix several warnings reported by clang (e.g. tautological comparisons)

+++ lib/sphere-pack/SpherePack.cpp
@@ -533,7 +533,7 @@ int SpherePack::addShadows(){
 
 int SpherePack::removeShadows(){
 	int ret=0;
-	for(size_t i=pack.size(); i>=0; i--){
+	for(long i=(long)pack.size(); i>=0; i--){
 		if(pack[i].shadowOf>=0){ pack.erase(pack.begin()+i); ret++; }
 	}
 	return ret;
